Renames and consolidates extension files for better understanding
